+++ pipeline/Dockerfile
@@ -11,6 +11,7 @@ COPY requirements.txt .
 RUN pip install -r requirements.txt
 
 COPY extract.py .
+COPY transform.py .
 COPY load.py .
 COPY rds_to_s3.py .
 COPY pipeline.py .
